fix issue with circular dependency on clean builds
