+++ fixtures/bbs-keygen.mjs
@@ -3,12 +3,13 @@
 // https://www.ietf.org/archive/id/draft-ietf-cose-bls-key-representations-02.html
 import {bbs} from "@mattrglobal/pairing-crypto"
 import {encode} from "jose/util/base64url"
-import {lineWrap} from "./bbs-fixtures.mjs"
-import fs from "fs/promises";
+import {lineWrap} from "./linewrap.mjs"
+import fs from "node:fs/promises";
 
 var keys = await bbs.bls12381_sha256.generateKeyPair();
 
 // create "build" directory if doesn't exist
+
 try { await fs.mkdir("build");  } catch (e) { /* ignore */ }
 
 var privateKeyStr = 
@@ -19,7 +20,8 @@ JSON.stringify({
     crv: "BLs12381G2",
     x: encode(keys.publicKey),
     d: encode(keys.secretKey)
-}, null, 2)
+}, null, 2);
+
 await fs.writeFile("build/private-key.jwk", privateKeyStr);
 await fs.writeFile("build/private-key.jwk.wrapped", lineWrap(privateKeyStr, 8));
 
@@ -31,5 +33,6 @@ JSON.stringify({
     crv: "BLs12381G2",
     x: encode(keys.publicKey)
 }, null, 2);
+
 await fs.writeFile("build/public-key.jwk", publicKeyStr);
 await fs.writeFile("build/public-key.jwk.wrapped", lineWrap(publicKeyStr, 8));
